+++ knode/utilities.cpp
@@ -296,7 +296,7 @@ QString KNHelper::rot13(const QString &s)
 {
   QString r(s);
 
-  for (int i=0; (uint)i<r.length(); i++) {
+  for (int i=0; i<r.length(); i++) {
     if ( r[i] >= QChar('A') && r[i] <= QChar('M') ||
          r[i] >= QChar('a') && r[i] <= QChar('m') )
       r[i] = (char)((int)QChar(r[i]).toLatin1() + 13);
